Split the code into smaller, reusable functions to improve readability and maintainability.
