+++ index.html
@@ -56,6 +56,7 @@
         </div>
     </div>
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/js/bootstrap.bundle.min.js"></script>
-    <script src="script.js"></script>
+    <script type="module" src="script.js"></script>
+    <script type="module" src="scripts/setModalProps.mjs"></script>
 </body>
 </html>
