+++ package.json
@@ -1,5 +1,8 @@
 {
   "dependencies": {
     "bootstrap": "^5.3.3"
+  },
+  "devDependencies": {
+    "jest": "^29.7.0"
   }
 }
