+++ scripts/chekcUserResponse.mjs
@@ -1,0 +1,4 @@
+let chekcUserResponse = (userAnswer) => {
+   return userAnswer == setAnswerCart() ? true : false
+}
+export default chekcUserResponse;
