+++ scripts/data.mjs
@@ -1,0 +1,7 @@
+let cartesShapeObject = {
+   1: 'Moon',
+   2: 'Wind',
+   3: 'Snow',
+   4: 'Nuages'
+}
+export default cartesShapeObject;
