+++ scripts/generateCartesShape.mjs
@@ -1,0 +1,7 @@
+function generateCartesShape(cartesTypes, cartesShapeObject) {
+  return cartesTypes.map(type => {
+      return cartesShapeObject[type]
+   })   
+
+}
+export default generateCartesShape;
