+++ scripts/processOperation.mjs
@@ -1,0 +1,6 @@
+import applyCardStyling from './applyCardStyling.mjs';
+import chekcUserResponse from './chekcUserResponse.mjs';
+let processOperation = (userAnswer, img) => {
+   applyCardStyling( chekcUserResponse(userAnswer), img)
+}   
+export default processOperation;
