+++ scripts/setAnswer.mjs
@@ -1,0 +1,4 @@
+import { cartesShape, cartesNumbers } from './cartes.js'
+let setAnswerCart = () => {
+   return cartesShape[cartesNumbers.indexOf(0)]
+}
