+++ style.css
@@ -54,14 +54,7 @@ body {
   75% { transform: translateX(8px); }
 }
 
-.modal-header {
-  border-bottom: 1px solid rgba(255, 255, 255, 0.1);
-}
 
-.modal-body {
-  font-weight: 500;
-  letter-spacing: 0.01em;
-}
 
 @media (max-width: 768px) {
   .wrapper {
