+++ todo
@@ -1,0 +1,23 @@
+Tips for Improvement
+Code Organization:
+
+Split the code into smaller, reusable functions to improve readability and maintainability.
+Consider using ES6 modules to separate concerns (e.g., random number generation, DOM manipulation, event handling).
+Error Handling:
+
+Add more robust error handling, especially for DOM manipulations and user interactions.
+Performance:
+
+Optimize DOM manipulations by using DocumentFragment or batching updates to reduce reflows and repaints.
+Accessibility:
+
+Ensure the modal and other interactive elements are accessible (e.g., focus management, ARIA roles).
+Styling:
+
+Use CSS classes for styling instead of inline styles to keep the JavaScript code clean.
+Code Comments:
+
+Add comments to explain the purpose of each function and complex logic.
+Testing:
+
+Write unit tests for the functions to ensure they work as expected
